Ver y crear tareas y triggers

+++ back_end/API/api.py
@@ -1,16 +1,17 @@
+from django.shortcuts import get_object_or_404
 from rest_framework.response import Response
-from .models import Persona, Usuario, Registro, Proyecto, Tarea, Recurso, Estado, Tipo, Prioridad, UsuariosRegistro, RecursoProyecto, RecursoTarea, UsuarioTarea, GroupUsuarioProyecto
-from .serializers import  RegistroUsuarioSerializer, PersonaSerializer, UsuarioSerializer, RegistroSerializer, GroupSerializer, ProyectoSerializer, TareaSerializer, RecursoSerializer, EstadoSerializer, TipoSerializer, PrioridadSerializer, UsuariosRegistroSerializer, RecursoProyectoSerializer, RecursoTareaSerializer, UsuarioTareaSerializer, GroupUsuarioProyectoSerializer
+from .models import (Persona, Usuario, Registro, Proyecto, Tarea, Recurso, Estado, Tipo, Prioridad, 
+                     UsuariosRegistro, RecursoProyecto, RecursoTarea, UsuarioTarea, GroupUsuarioProyecto)
+from .serializers import (RegistroUsuarioSerializer, PersonaSerializer, UsuarioSerializer, RegistroSerializer, 
+                          GroupSerializer, ProyectoSerializer, TareaSerializer, RecursoSerializer, EstadoSerializer, 
+                          TipoSerializer, PrioridadSerializer, UsuariosRegistroSerializer, RecursoProyectoSerializer, 
+                          RecursoTareaSerializer, UsuarioTareaSerializer, GroupUsuarioProyectoSerializer)
 from rest_framework import viewsets, generics
 from rest_framework.permissions import IsAuthenticated, DjangoModelPermissions, AllowAny
 from rest_framework.exceptions import PermissionDenied
 from django.contrib.auth.models import Group
-from rest_framework import viewsets, status
-from django.contrib.auth.models import Group
 from rest_framework.decorators import api_view, authentication_classes, permission_classes
-from rest_framework.response import Response
 from rest_framework import status
-from .serializers import RegistroSerializer
 
 @api_view(['POST'])
 @authentication_classes([])
@@ -30,8 +31,8 @@ class UsuarioViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         usuario = self.request.user
-        user = Usuario.objects.filter(id=usuario.id)
-        return user
+        return Usuario.objects.filter(id=usuario.id)
+
 
 class ProyectoViewSet(viewsets.ModelViewSet):
     queryset = Proyecto.objects.all()
@@ -40,8 +41,7 @@ class ProyectoViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         usuario = self.request.user
-        proyectos_usuario = Proyecto.objects.filter(groupusuarioproyecto__idUsuario=usuario)
-        return proyectos_usuario
+        return Proyecto.objects.filter(groupusuarioproyecto__idUsuario=usuario)
 
     def perform_create(self, serializer):
         proyecto = serializer.save()
@@ -64,6 +64,7 @@ class ProyectoViewSet(viewsets.ModelViewSet):
         else:
             raise PermissionDenied('No tienes permiso para eliminar este proyecto')
 
+
 class TareaViewSet(viewsets.ModelViewSet):
     queryset = Tarea.objects.all()
     permission_classes = [IsAuthenticated]
@@ -72,30 +73,43 @@ class TareaViewSet(viewsets.ModelViewSet):
     def get_queryset(self):
         usuario = self.request.user
         proyecto_id = self.kwargs.get('proyecto_id')
-        
+    
         if proyecto_id:
-            try:
-                proyecto = Proyecto.objects.get(id=proyecto_id, groupusuarioproyecto__idUsuario=usuario)
-                return Tarea.objects.filter(idProyecto=proyecto)
-            except Proyecto.DoesNotExist:
-                return Tarea.objects.none()
+            proyectos_usuario = Proyecto.objects.filter(groupusuarioproyecto__idUsuario=usuario, id=proyecto_id)
+            return Tarea.objects.filter(idProyecto__in=proyectos_usuario)
         else:
             proyectos_usuario = Proyecto.objects.filter(groupusuarioproyecto__idUsuario=usuario)
             return Tarea.objects.filter(idProyecto__in=proyectos_usuario)
-    
 
     def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        self.perform_create(serializer)
-        headers = self.get_success_headers(serializer.data)
-        return Response(serializer.data, status=201, headers=headers)
-            
+        usuario_actual = request.user
+        proyecto = get_object_or_404(Proyecto, id=request.data.get('idProyecto'))
+
+        if not GroupUsuarioProyecto.objects.filter(
+            idUsuario=usuario_actual,
+            idProyecto=proyecto,
+            idGroup__permissions__codename='add_tarea'
+        ).exists():
+            return Response({'error': 'No tienes permiso para agregar tareas a este proyecto'}, status=status.HTTP_403_FORBIDDEN)
+
+        return super().create(request, *args, **kwargs)
+
     def perform_update(self, serializer):
-        serializer.save()
+        tarea = self.get_object()
+        usuario = self.request.user
+        if usuario.has_perm('app_name.change_tarea', tarea):
+            serializer.save()
+        else:
+            raise PermissionDenied('No tienes permiso para editar esta tarea')
 
     def perform_destroy(self, instance):
-        instance.delete()
+        usuario = self.request.user
+        if usuario.has_perm('app_name.delete_tarea', instance):
+            instance.delete()
+        else:
+            raise PermissionDenied('No tienes permiso para eliminar esta tarea')
+
+
 class GroupUsuarioProyectoViewSet(viewsets.ModelViewSet):
     queryset = GroupUsuarioProyecto.objects.all()
     permission_classes = [IsAuthenticated]
@@ -119,24 +133,22 @@ class GroupUsuarioProyectoViewSet(viewsets.ModelViewSet):
         except (Proyecto.DoesNotExist, Usuario.DoesNotExist, Group.DoesNotExist):
             return Response({'error': 'Proyecto, usuario o rol no encontrado'}, status=status.HTTP_400_BAD_REQUEST)
 
-        print(GroupUsuarioProyecto.objects.filter(idUsuario_id__id=usuario_actual.id, idProyecto_id__id=proyecto.id))
-
         if GroupUsuarioProyecto.objects.filter(
-            idUsuario_id__id=usuario_actual.id,
-            idProyecto_id__id=proyecto.id,
-            idGroup_id__permissions__codename='api.add_groupusuarioproyecto'
+            idUsuario=usuario_actual,
+            idProyecto=proyecto,
+            idGroup__permissions__codename='add_groupusuarioproyecto'
         ).exists():
             GroupUsuarioProyecto.objects.create(idGroup=group, idUsuario=usuario, idProyecto=proyecto)
             return Response({'message': 'Usuario agregado al proyecto correctamente'}, status=status.HTTP_201_CREATED)
         else:
             return Response({'error': 'No tienes permiso para agregar usuarios a este proyecto'}, status=status.HTTP_403_FORBIDDEN)
-    
+
     def perform_update(self, serializer):
         group_usuario_proyecto = self.get_object()
         proyecto = group_usuario_proyecto.idProyecto
         usuario_actual = self.request.user
 
-        if usuario_actual.has_perm('API.change_groupusuarioproyecto', proyecto):
+        if usuario_actual.has_perm('app_name.change_groupusuarioproyecto', proyecto):
             serializer.save()
         else:
             raise PermissionDenied('No tienes permiso para editar roles en este proyecto')
@@ -149,6 +161,7 @@ class GroupUsuarioProyectoViewSet(viewsets.ModelViewSet):
             instance.delete()
         else:
             raise PermissionDenied('No tienes permiso para eliminar roles en este proyecto')
+
 
 class RegistroUsuarioView(generics.CreateAPIView):
     queryset = Usuario.objects.all()
